Tested whether the total cost was updated
There was an issue where the total cost was not updated for some reason. (When adding a new menu item, the cost was not added to the total cost, and the total cost remained the cost of the first menu item added.)

Method: I defined the totalCost variable as a reactive assignment, which would automatically recalculate totalCost every time selectedItems was updated, so that when I added or removed new items the total cost would be updated immediately.

Result: Instead of the total cost being added up, the selected items (and total cost) themselves were no longer displayed.