Tested whether the total cost was updated (2)
Method: Even though totalCost was calculated reactively using $:, I was calling it as a normal function, totalCost(), so I fixed that and retested.

Result: Although totalCost is calculated reactively using $:, it is not a normal function call, so it was a mistake to call it like totalCost().
The issue of the total cost not being added up and the issue of the selected items not being displayed that occurred when trying to fix that have both been resolved.