Added message on properly timing.
I have added alert messages to display when data is loading (Loading...) and when there are no items in the menu (when a user tries to add the same menu multiple times) (You've already selected the item).

Test 1: Then, I tested pressing the Add to Menu button several times.

Boundary: Press the button several times and select the selected item again

Result 1: As expected, even if I tried to select it multiple times, I could only select it once. (I pressed the button several times in succession and confirmed that the pop-up message "You've already selected the item. You can only select once." was displayed each time, and couldn't add the item to Selected Items section again. (The total amount also did not change.))

Test 2: I also tested by refreshing the page several times (in a place in the school where the internet speed was a little slower to take time to load content), and confirmed that the "Loading..." message was actually displayed under the event name.

Result 2: As expected, the message was displayed properly each time the page was refreshed and before the content was displayed.